+++ src/Allegro.CosmosDb.BatchUtilities/Allegro.CosmosDb.BatchUtilities/AutoScaler/ICosmosAutoScalerFactory.cs
@@ -1,0 +1,11 @@
+namespace Allegro.CosmosDb.BatchUtilities;
+
+public interface ICosmosAutoScalerFactory
+{
+    ICosmosAutoScaler ForContainer(
+        string databaseName,
+        string containerName);
+
+    ICosmosAutoScaler ForDatabase(
+        string databaseName);
+}