+++ src/Allegro.CosmosDb.BatchUtilities/Allegro.CosmosDb.BatchUtilities/AutoScaler/ICosmosAutoScalerFactoryProvider.cs
@@ -1,0 +1,7 @@
+namespace Allegro.CosmosDb.BatchUtilities;
+
+public interface ICosmosAutoScalerFactoryProvider
+{
+    public ICosmosAutoScalerFactory GetFactory(
+        string clientName);
+}