+++ src/Allegro.CosmosDb.BatchUtilities/Allegro.CosmosDb.BatchUtilities/BatchClient/ICosmosBatchClientProvider.cs
@@ -1,0 +1,8 @@
+using Microsoft.Azure.Cosmos;
+
+namespace Allegro.CosmosDb.BatchUtilities;
+
+public interface ICosmosBatchClientProvider
+{
+    CosmosClient GetBatchClient(string clientName);
+}