+++ src/Allegro.CosmosDb.BatchUtilities/Allegro.CosmosDb.BatchUtilities/Extensions/CosmosExtensions.cs
@@ -3,21 +3,24 @@ using System.Collections.Generic;
 using System.Linq;
 using Allegro.CosmosDb.BatchUtilities.Configuration;
 using Allegro.CosmosDb.BatchUtilities.Events;
-using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Fluent;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 
 namespace Allegro.CosmosDb.BatchUtilities.Extensions
 {
+    public record BatchClientServiceProviderMarker(string ClientName);
+
     public static class CosmosExtensions
     {
-        public static IServiceCollection AddCosmosBatchUtilitiesFromConfiguration(
+        public static IServiceCollection AddCosmosBatchClientFromConfiguration(
             this IServiceCollection services,
+            Func<IServiceProvider, CosmosClientBuilder> cosmosClientBuilderFunc,
             IConfiguration configuration,
             string configurationSection,
-            Action<CosmosClientBuilder>? additionalCosmosConfiguration = null)
+            string clientName)
         {
             var configurations = configuration
                 .GetSection(configurationSection)
@@ -47,69 +50,76 @@ namespace Allegro.CosmosDb.BatchUtilities.Extensions
                 }
             }
 
-            return services.AddCosmosBatchUtilities(
-                additionalCosmosConfiguration,
-                registrations.ToArray());
+            return services.AddCosmosBatchClient(
+                cosmosClientBuilderFunc,
+                clientName,
+                registrations.Select(p => new Func<IServiceProvider, BatchUtilitiesRegistration>(_ => p)).ToArray());
         }
 
-        public static IServiceCollection AddCosmosBatchUtilities(
+        public static IServiceCollection AddCosmosBatchClient(
             this IServiceCollection services,
-            params BatchUtilitiesRegistration[] registrations)
+            Func<IServiceProvider, CosmosClientBuilder> cosmosClientBuilderFunc,
+            string clientName,
+            params BatchUtilitiesRegistration[] registrationFactories)
         {
-            return services.AddCosmosBatchUtilities(null, registrations);
-        }
-
-        public static IServiceCollection AddCosmosBatchUtilities(
-            this IServiceCollection services,
-            Action<CosmosClientBuilder>? additionalCosmosConfiguration,
-            params BatchUtilitiesRegistration[] registrations)
-        {
-            return services.AddCosmosBatchUtilities(
-                (builder, _) => additionalCosmosConfiguration?.Invoke(builder),
-                registrations
-                    .Select(x => new Func<IServiceProvider, BatchUtilitiesRegistration>(_ => x))
+            services.AddCosmosBatchClient(
+                cosmosClientBuilderFunc,
+                clientName,
+                registrationFactories.Select(p => new Func<IServiceProvider, BatchUtilitiesRegistration>(_ => p))
                     .ToArray());
+
+            return services;
         }
 
-        public static IServiceCollection AddCosmosBatchUtilities(
+        public static IServiceCollection AddCosmosBatchClient(
             this IServiceCollection services,
-            Action<CosmosClientBuilder, IServiceProvider>? additionalCosmosConfiguration,
+            Func<IServiceProvider, CosmosClientBuilder> cosmosClientBuilderFunc,
+            string clientName,
             params Func<IServiceProvider, BatchUtilitiesRegistration>[] registrationFactories)
         {
-            services.AddSingleton(sp => registrationFactories.Select(f => f(sp)));
-            services.AddSingleton<ICosmosClientProvider>(
-                sp =>
-                {
-                    var registrations = sp.GetRequiredService<IEnumerable<BatchUtilitiesRegistration>>();
-                    var cosmosClient = sp.GetRequiredService<CosmosClient>();
-                    var builderFactory = sp.GetRequiredService<Func<CosmosClientBuilder>>();
-                    var logger = sp.GetRequiredService<ILogger<BatchRequestHandler>>();
+            var name = clientName.ToLowerInvariant();
 
-                    var batchClientBuilder = builderFactory()
-                        .AddCustomHandlers(new BatchRequestHandler(logger, registrations.ToArray()));
+            services.TryAddSingleton<ICosmosBatchClientProvider, CosmosBatchClientProvider>();
+            services.TryAddSingleton<ICosmosAutoScalerFactoryProvider, CosmosAutoScalerFactoryProvider>();
 
-                    additionalCosmosConfiguration?.Invoke(batchClientBuilder, sp);
+            if (!services.Any(
+                    p => p.ServiceType == typeof(BatchClientServiceProviderMarker) &&
+                         ((BatchClientServiceProviderMarker)p.ImplementationInstance!).ClientName == name))
+            {
+                services.AddSingleton(new BatchClientServiceProviderMarker(name));
+                services.AddSingleton<CosmosBatchClientBuilderWrapper>(
+                    sp =>
+                    {
+                        var registrations = registrationFactories.Select(f => f(sp));
+                        var logger = sp.GetRequiredService<ILogger<BatchRequestHandler>>();
+                        var batchClientBuilder = cosmosClientBuilderFunc(sp)
+                            .AddCustomHandlers(new BatchRequestHandler(logger, registrations.ToArray()));
 
-                    return new CosmosClientProvider(
-                        cosmosClient,
-                        batchClientBuilder.Build());
-                });
+                        return new CosmosBatchClientBuilderWrapper(name, batchClientBuilder);
+                    });
 
-            services.AddHostedService(sp => new CosmosAutoScalerInitializationHostedService(sp));
-            services.AddSingleton<ICosmosAutoScalerFactory>(
-                sp =>
-                {
-                    var registrations = sp.GetRequiredService<IEnumerable<BatchUtilitiesRegistration>>()
-                        .Where(x => x.CosmosAutoScalerConfiguration?.Enabled == true);
-                    var cosmosClient = sp.GetRequiredService<CosmosClient>();
-                    var loggerFactory = sp.GetRequiredService<ILoggerFactory>();
-                    var cosmosMetricsEventHandlers = sp.GetServices<CosmosAutoScalerMetricsCalculatedEventHandler>();
-                    return new CosmosAutoScalerFactory(
-                        cosmosClient,
-                        loggerFactory,
-                        cosmosMetricsEventHandlers,
-                        registrations.ToArray());
-                });
+                services.AddSingleton<CosmosAutoScalerFactoryWrapper>(
+                    sp =>
+                    {
+                        var registrations = registrationFactories.Select(f => f(sp))
+                            .Where(x => x.CosmosAutoScalerConfiguration?.Enabled == true);
+                        var cosmosClientBatchProvider = sp.GetRequiredService<ICosmosBatchClientProvider>();
+                        var loggerFactory = sp.GetRequiredService<ILoggerFactory>();
+                        var cosmosMetricsEventHandlers =
+                            sp.GetServices<CosmosAutoScalerMetricsCalculatedEventHandler>();
+                        var factory = new CosmosAutoScalerFactory(
+                            cosmosClientBatchProvider.GetBatchClient(name),
+                            loggerFactory,
+                            cosmosMetricsEventHandlers,
+                            registrations.ToArray());
+                        var factoryWrapper = new CosmosAutoScalerFactoryWrapper(name, factory);
+                        return factoryWrapper;
+                    });
+            }
+            else
+            {
+                throw new InvalidOperationException($"Cosmos batch client with name = '{name}' is already registered!");
+            }
 
             return services;
         }